+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 1.0.0+2025-01-23
+- Add UMAP and PCA options for high-dimensional projections in `analysis`
+- Bugfix for infinitely scaling discrepancy in `ViewTemporalDiscrepancy`
+- Compatibility with reduced states in `View3D`
+- Fix incorrect identification of `modal_targets` using `base_env` in analysis
+- Formatting for `tqdm` in `analysis`, especially for notebooks
+- Lower tensor precision inplace for saving memories
+- Properly assign number of cell pairings to temporal analysis
+- Refresh runs
+- Undo model stage indexing change
+
 ### 1.0.0+2025-01-22 (1-3)
 - Add `View3D` compatibility with 4+ dimensional state input
 - Add alternative name to `MMD-MA` dataset, `MMDMA`